requirements.txt
pandas
pydantic
openai
streamlit
python-dotenv
matplotlib
loguru